+++ test/fixtures/parser/ansi/select_from_where_overlaps.sql
@@ -1,0 +1,1 @@
+SELECT column_name FROM table_name WHERE period1 OVERLAPS period2;